Move About section styling to individual jsx files
