+++ src/components/About/About.js
@@ -1,12 +1,16 @@
 import React from "react"
 
-import AboutImg from "../assets/images/designs/about.svg"
-import * as Styles from "../styles/components/index.js"
+import Wrapper from "./Wrapper"
+import Content from "./Content"
+import Text from "./Text"
+import Image from "./Image"
+
+import AboutImg from "../../assets/images/designs/about.svg"
 
 const About = ({ scrollRef }) => (
-  <Styles.About ref={scrollRef}>
-    <Styles.AboutContent>
-      <Styles.AboutText>
+  <Wrapper ref={scrollRef}>
+    <Content>
+      <Text>
         <h2>Behind Canada’s original collegiate hackathon</h2>
         <p>
           McHacks isn't your typical hackathon, and our hackers aren't either!
@@ -20,10 +24,10 @@ const About = ({ scrollRef }) => (
           for your next big hack. We hope you'll join us this winter as we jump
           into another round of Canada's favourite hackathon!
         </p>
-      </Styles.AboutText>
-      <Styles.AboutImg src={AboutImg} />
-    </Styles.AboutContent>
-  </Styles.About>
+      </Text>
+      <Image src={AboutImg} />
+    </Content>
+  </Wrapper>
 )
 
 export default About
